requirements.txt
beautifulsoup4==4.7.1
boto3==1.9.179
botocore==1.12.179
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-crispy-forms==1.7.2
django-endless-pagination==2.0
django-extensions==2.1.6
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-storages==1.7.1
django-unused-media==0.1.13
docutils==0.14
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
jdcal==1.4
jmespath==0.9.4
jsonfield==2.0.2
openpyxl==2.6.1
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9
urllib3==1.24.1